+++ Credixware/Internal/SDK/IClientThinkable.h
@@ -53,9 +53,9 @@ class IClientUnknown : public IHandleEntity
 {
 public:
 	virtual ICollideable*		GetCollideable() = 0;
-	virtual IClientNetworkable*	GetClientNetworkable() = 0;
-	virtual IClientRenderable*	GetClientRenderable() = 0;
-	virtual IClientEntity*		GetIClientEntity() = 0;
+	virtual class IClientNetworkable*	GetClientNetworkable() = 0;
+	virtual class IClientRenderable*	GetClientRenderable() = 0;
+	virtual class IClientEntity*		GetIClientEntity() = 0;
 	virtual C_BaseEntity*		GetBaseEntity() = 0;
 	virtual IClientThinkable*	GetClientThinkable() = 0;
 };